remove main from logger and write test function
